backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
psycopg[binary]==3.2.9
python-dotenv==1.0.1
pydantic>=2.9,<3
drf-spectacular==0.27.2
django-cors-headers==4.4.0
pytest==8.2.0
pytest-django==4.9.0
requests==2.32.3
